requirements.txt
# Core dependencies
azure-ai-agent-framework>=0.1.0
azure-ai-inference>=1.0.0
azure-identity>=1.15.0
azure-monitor-opentelemetry>=1.2.0

# Azure SDK
azure-mgmt-compute>=30.0.0
azure-mgmt-network>=25.0.0
azure-keyvault-secrets>=4.7.0
azure-search-documents>=11.4.0

# Microsoft Graph
msgraph-sdk>=1.0.0
msal>=1.24.0

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP clients
httpx>=0.26.0
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.3
pytz>=2023.3

# Logging and monitoring
opencensus-ext-azure>=1.1.13
python-json-logger>=2.0.7

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code quality
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0
